Changed MemoryBlockSouceInfo into an interface
