+++ Ghidra/Framework/SoftwareModeling/src/main/java/ghidra/program/database/mem/AddressSourceInfo.java
@@ -21,8 +21,7 @@ import java.util.Optional;
 
 import ghidra.program.model.address.Address;
 import ghidra.program.model.address.AddressRange;
-import ghidra.program.model.mem.Memory;
-import ghidra.program.model.mem.MemoryBlock;
+import ghidra.program.model.mem.*;
 
 /**
  * Provides information about the source of a byte value at an address including the file it
