+++ Ghidra/Framework/SoftwareModeling/src/main/java/ghidra/program/model/mem/MemoryBlock.java
@@ -20,7 +20,6 @@ import java.io.Serializable;
 import java.util.List;
 
 import ghidra.framework.store.LockException;
-import ghidra.program.database.mem.MemoryBlockSourceInfo;
 import ghidra.program.model.address.Address;
 import ghidra.program.model.listing.Program;
 import ghidra.util.exception.DuplicateNameException;
